+++ src/x3f_matrix.h
@@ -1,21 +1,21 @@
-extern void x3f_scalar_3x1_mul(float a, float *b, float *c);
-extern void x3f_scalar_3x3_mul(float a, float *b, float *c);
+extern void x3f_scalar_3x1_mul(double a, double *b, double *c);
+extern void x3f_scalar_3x3_mul(double a, double *b, double *c);
 
-extern void x3f_3x3_3x1_mul(float *a, float *b, float *c);
-extern void x3f_3x3_3x3_mul(float *a, float *b, float *c);
+extern void x3f_3x3_3x1_mul(double *a, double *b, double *c);
+extern void x3f_3x3_3x3_mul(double *a, double *b, double *c);
 
-extern void x3f_3x3_diag(float *a, float *b);
+extern void x3f_3x3_diag(double *a, double *b);
 
-extern void x3f_3x1_print(float *a);
-extern void x3f_3x3_print(float *a);
+extern void x3f_3x1_print(double *a);
+extern void x3f_3x3_print(double *a);
 
-extern void x3f_XYZ_to_ProPhotoRGB(float *a);
+extern void x3f_XYZ_to_ProPhotoRGB(double *a);
 
-extern void x3f_XYZ_to_AdobeRGB(float *a);
-extern void x3f_AdobeRGB_to_XYZ(float *a);
+extern void x3f_XYZ_to_AdobeRGB(double *a);
+extern void x3f_AdobeRGB_to_XYZ(double *a);
 
-extern void x3f_XYZ_to_sRGB(float *a);
+extern void x3f_XYZ_to_sRGB(double *a);
 
-extern void x3f_CIERGB_to_XYZ(float *a);
+extern void x3f_CIERGB_to_XYZ(double *a);
 
-extern void x3f_Bradford_D50_to_D65(float *a);
+extern void x3f_Bradford_D50_to_D65(double *a);
